+++ lib/active_model_serializers/adapter/json_api.rb
@@ -8,11 +8,13 @@ module ActiveModelSerializers
       require 'active_model/serializer/adapter/json_api/meta'
       autoload :Deserialization
       require 'active_model/serializer/adapter/json_api/api_objects'
+      autoload :Error
 
       # TODO: if we like this abstraction and other API objects to it,
       # then extract to its own file and require it.
       module ApiObjects
-        module JsonApi
+        # {http://jsonapi.org/format/#document-jsonapi-object Jsonapi Object}
+        module Jsonapi
           ActiveModelSerializers.config.jsonapi_version = '1.0'
           ActiveModelSerializers.config.jsonapi_toplevel_meta = {}
           # Make JSON API top-level jsonapi member opt-in
@@ -50,9 +52,19 @@ module ActiveModelSerializers
         @fieldset = options[:fieldset] || ActiveModel::Serializer::Fieldset.new(options.delete(:fields))
       end
 
+      # {http://jsonapi.org/format/#crud Requests are transactional, i.e. success or failure}
+      # {http://jsonapi.org/format/#document-top-level data and errors MUST NOT coexist in the same document.}
       def serializable_hash(options = nil)
         options ||= {}
+        if serializer.success?
+          success_document(options)
+        else
+          failure_document
+        end
+      end
 
+      # {http://jsonapi.org/format/#document-top-level Primary data}
+      def success_document(options)
         is_collection = serializer.respond_to?(:each)
         serializers = is_collection ? serializer : [serializer]
         primary_data, included = resource_objects_for(serializers)
@@ -61,7 +73,7 @@ module ActiveModelSerializers
         hash[:data] = is_collection ? primary_data : primary_data[0]
         hash[:included] = included if included.any?
 
-        ApiObjects::JsonApi.add!(hash)
+        ApiObjects::Jsonapi.add!(hash)
 
         if instance_options[:links]
           hash[:links] ||= {}
@@ -76,6 +88,29 @@ module ActiveModelSerializers
         hash
       end
 
+      # {http://jsonapi.org/format/#errors JSON API Errors}
+      # TODO: look into caching
+      # rubocop:disable Style/AsciiComments
+      # definition:
+      #   ☑ toplevel_errors array (required)
+      #   ☐ toplevel_meta
+      #   ☐ toplevel_jsonapi
+      # rubocop:enable Style/AsciiComments
+      def failure_document
+        hash = {}
+        # PR Please :)
+        # ApiObjects::Jsonapi.add!(hash)
+
+        if serializer.respond_to?(:each)
+          hash[:errors] = serializer.flat_map do |error_serializer|
+            Error.resource_errors(error_serializer)
+          end
+        else
+          hash[:errors] = Error.resource_errors(serializer)
+        end
+        hash
+      end
+
       def fragment_cache(cached_hash, non_cached_hash)
         root = false if instance_options.include?(:include)
         ActiveModelSerializers::Adapter::JsonApi::FragmentCache.new.fragment_cache(root, cached_hash, non_cached_hash)
@@ -87,6 +122,7 @@ module ActiveModelSerializers
 
       private
 
+      # {http://jsonapi.org/format/#document-resource-objects Primary data}
       def resource_objects_for(serializers)
         @primary = []
         @included = []
@@ -128,10 +164,12 @@ module ActiveModelSerializers
         process_relationships(serializer, include_tree)
       end
 
+      # {http://jsonapi.org/format/#document-resource-object-attributes Document Resource Object Attributes}
       def attributes_for(serializer, fields)
         serializer.attributes(fields).except(:id)
       end
 
+      # {http://jsonapi.org/format/#document-resource-objects Document Resource Objects}
       def resource_object_for(serializer)
         resource_object = cache_check(serializer) do
           resource_object = ActiveModel::Serializer::Adapter::JsonApi::ApiObjects::ResourceIdentifier.new(serializer).as_json
@@ -155,6 +193,7 @@ module ActiveModelSerializers
         resource_object
       end
 
+      # {http://jsonapi.org/format/#document-resource-object-relationships Document Resource Object Relationship}
       def relationships_for(serializer, requested_associations)
         include_tree = ActiveModel::Serializer::IncludeTree.from_include_args(requested_associations)
         serializer.associations(include_tree).each_with_object({}) do |association, hash|
@@ -168,16 +207,19 @@ module ActiveModelSerializers
         end
       end
 
+      # {http://jsonapi.org/format/#document-links Document Links}
       def links_for(serializer)
         serializer._links.each_with_object({}) do |(name, value), hash|
           hash[name] = Link.new(serializer, value).as_json
         end
       end
 
+      # {http://jsonapi.org/format/#fetching-pagination Pagination Links}
       def pagination_links_for(serializer, options)
         JsonApi::PaginationLinks.new(serializer.object, options[:serialization_context]).serializable_hash(options)
       end
 
+      # {http://jsonapi.org/format/#document-meta Docment Meta}
       def meta_for(serializer)
         ActiveModel::Serializer::Adapter::JsonApi::Meta.new(serializer).as_json
       end